# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: MyASA 2.0 API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /show-books/{id}/blocks:
    post:
      operationId: createShowBookBlock
      tags: [show-book]
      summary: Adicionar bloco ao livro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockCreate"
      responses:
        "201":
          description: Bloco criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  block:
                    $ref: "#/components/schemas/ShowBookBlock"
                required: [block]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /show-books/{id}/blocks/{blockId}:
    patch:
      operationId: updateShowBookBlock
      tags: [show-book]
      summary: Atualizar bloco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: blockId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUpdate"
      responses:
        "200":
          description: Bloco atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  block:
                    $ref: "#/components/schemas/ShowBookBlock"
                required: [block]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteShowBookBlock
      tags: [show-book]
      summary: Remover bloco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: blockId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionalReasonPayload"
      responses:
        "204":
          description: Bloco removido
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /daily-book/{id}/blocks/{blockId}:
    delete:
      operationId: deleteDailyBookBlock
      tags: [daily-book]
      summary: "Remover bloco do Livro do Dia (soft delete)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: blockId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bloco removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  block:
                    $ref: "#/components/schemas/DailyBookBlock"
                required: [block]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ShowBookBlock:
      type: object
      properties:
        id:
          type: string
        showBookId:
          type: string
        sceneId:
          type: string
          nullable: true
        name:
          type: string
        order:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, showBookId, name, order]
...
    BlockCreate:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        sceneId:
          type: string
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        reason:
          type: string
      required: [name, order]
    BlockUpdate:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        changeType:
          type: string
          enum: [STRUCTURAL, CONFIG]
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        reason:
          type: string
      required: []
...
    OptionalReasonPayload:
      type: object
      properties:
        reason:
          type: string
      required: []

    # ── Agenda — Schemas ──────────────────────────────────────────────────────
...
    DailyBookBlock:
      type: object
      properties:
        id:
          type: string
        dailyBookId:
          type: string
        sceneId:
          type: string
          nullable: true
        name:
          type: string
        order:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        sourceBlockId:
          type: string
          nullable: true
        isRemoved:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, dailyBookId, name, order, isRemoved, createdAt, updatedAt]
...